requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
Werkzeug==3.0.4
python-dotenv==1.0.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
Flask-Bcrypt==1.0.1
email-validator==2.1.1
Flask-CORS==4.0.0
python-dateutil==2.9.0